DimHierarchyLevel repository, service, controller done

+++ src/MarketStat.Database/MarketStat.Database.Repositories/PostgresRepositories/Dimensions/DimHierarchyLevelRepository.cs
@@ -1,9 +1,11 @@
 using MarketStat.Common.Converter.MarketStat.Common.Converter.Dimensions;
 using MarketStat.Common.Core.MarketStat.Common.Core.Dimensions;
+using MarketStat.Common.Exceptions;
 using MarketStat.Database.Context;
 using MarketStat.Database.Core.Repositories.Dimensions;
 using MarketStat.Database.Models;
 using Microsoft.EntityFrameworkCore;
+using Npgsql;
 
 namespace MarketStat.Database.Repositories.PostgresRepositories.Dimensions;
 
@@ -23,7 +25,17 @@ public class DimHierarchyLevelRepository : BaseRepository, IDimHierarchyLevelRep
             hierarchyLevelName: dimHierarchyLevel.HierarchyLevelName
         );
         await _dbContext.DimHierarchyLevels.AddAsync(dbModel);
-        await _dbContext.SaveChangesAsync();
+        
+        try
+        {
+            await _dbContext.SaveChangesAsync();
+        }
+        catch (DbUpdateException dbEx)
+            when (dbEx.InnerException is PostgresException pg
+                  && pg.SqlState == PostgresErrorCodes.UniqueViolation)
+        {
+            throw new ConflictException($"A hierarchy level '{dimHierarchyLevel.HierarchyLevelName}' already exists.");
+        }
         dimHierarchyLevel.HierarchyLevelId = dbModel.HierarchyLevelId;
     }
 
@@ -35,24 +47,37 @@ public class DimHierarchyLevelRepository : BaseRepository, IDimHierarchyLevelRep
 
     public async Task<DimHierarchyLevel> GetHierarchyLevelByIdAsync(int id)
     {
-        var dbHierarchyLevel = await _dbContext.DimHierarchyLevels.FindAsync(id) 
-                               ?? throw new KeyNotFoundException($"HierarchyLevel {id} not found.");
+        var dbHierarchyLevel = await _dbContext.DimHierarchyLevels.FindAsync(id);
+        if (dbHierarchyLevel is null)
+            throw new NotFoundException($"Hierarchy level {id} not found.");
         return DimHierarchyLevelConverter.ToDomain(dbHierarchyLevel);
     }
 
     public async Task UpdateHierarchyLevelAsync(DimHierarchyLevel dimHierarchyLevel)
     {
-        var dbHierarchyLevel = await _dbContext.DimHierarchyLevels.FindAsync(dimHierarchyLevel.HierarchyLevelId)
-                               ?? throw new KeyNotFoundException(
-                                   $"Cannot update: HierarchyLevel {dimHierarchyLevel.HierarchyLevelId} not found.");
+        var dbHierarchyLevel = await _dbContext.DimHierarchyLevels.FindAsync(dimHierarchyLevel.HierarchyLevelId);
+        if (dbHierarchyLevel is null)
+            throw new NotFoundException($"Hierarchy level {dimHierarchyLevel.HierarchyLevelId} not found.");
+        
         dbHierarchyLevel.HierarchyLevelName = dimHierarchyLevel.HierarchyLevelName;
-        await _dbContext.SaveChangesAsync();
+        
+        try
+        {
+            await _dbContext.SaveChangesAsync();
+        }
+        catch (DbUpdateException dbEx)
+            when (dbEx.InnerException is PostgresException pg
+                  && pg.SqlState == PostgresErrorCodes.UniqueViolation)
+        {
+            throw new ConflictException($"A hierarchy level '{dimHierarchyLevel.HierarchyLevelName}' already exists.");
+        }
     }
 
     public async Task DeleteHierarchyLevelAsync(int id)
     {
-        var dbHierarchyLevel = await _dbContext.DimHierarchyLevels.FindAsync(id)
-                               ?? throw new KeyNotFoundException($"Cannot update: HierarchyLevel {id} not found.");
+        var dbHierarchyLevel = await _dbContext.DimHierarchyLevels.FindAsync(id);
+        if (dbHierarchyLevel is null)
+            throw new NotFoundException($"Hierarchy level {id} not found.");
         _dbContext.DimHierarchyLevels.Remove(dbHierarchyLevel);
         await _dbContext.SaveChangesAsync();
     }
